Bug 1035321 - Send the accuracy of locations to Find My Device's server.
This also imports mock_l10n.js from bug 1004973, since the tests introduced
in this patch depend on that file.
